move firebase and timer to app
